move all of the fixtures into a separate file to keep spec_helper cleaner
